Join the test files as one
